+++ .eslintrc.json
@@ -1,3 +1,3 @@
 {
-  "extends": "eslint-config-es/2015/server"
+  "extends": "es/node"
 }
